records

+++ Firmware/Core/Src/flash.c
@@ -53,7 +53,7 @@ uint32_t Flash_Write_Data(uint32_t StartPageAddress, uint64_t *Data, uint16_t nu
 void Flash_Read_Data(uint32_t StartPageAddress, uint64_t *RxBuf, uint16_t numberofdoublewords)
 {
     while (1) {
-        *RxBuf = *(__IO uint32_t *)StartPageAddress;
+        *RxBuf = *(__IO uint64_t *)StartPageAddress;
         StartPageAddress += 8;
         RxBuf++;
         if (!(numberofdoublewords--))
